# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Api",
        "version": "0.1"
    },
    "paths": {
        "/api/setting": {
            "patch": {
                "tags": [
                    "Setting Update"
                ],
                "summary": "Update Setting",
                "operationId": "5f0b646d948986e771a7d110c0f6c885",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "key": {
                                                "type": "string"
                                            },
                                            "value": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "key": "overtime_method",
                                    "value": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "key": {
                                            "type": "string",
                                            "example": "overtime_method"
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": 1
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/employees": {
            "post": {
                "tags": [
                    "Create Employees"
                ],
                "summary": "Create Employess",
                "operationId": "8adf618a69e276e5dffd7cf2e209fc2e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "salary": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "ahmad",
                                    "salary": 3000000
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "ahmad"
                                        },
                                        "salary": {
                                            "type": "integer",
                                            "example": 3000000
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/overtimes": {
            "post": {
                "tags": [
                    "Create Overtimes"
                ],
                "summary": "Create Overtimes",
                "operationId": "44e01e71c959b4355b08ae176aa6349f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "employees_id": {
                                                "type": "integer"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "time_started": {
                                                "type": "time"
                                            },
                                            "time_ended": {
                                                "type": "time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "employees_id": 1,
                                    "date": "2022-10-10",
                                    "time_started": "18:00",
                                    "time_ended": "20:30"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "employees_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-10-10"
                                        },
                                        "time_started": {
                                            "type": "time",
                                            "example": "18:00"
                                        },
                                        "time_ended": {
                                            "type": "time",
                                            "example": "20:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/overtimes-pay/calculate/{date}": {
            "get": {
                "tags": [
                    "Overtimes Calculate"
                ],
                "summary": "Get Overtimes Calculate",
                "operationId": "d30a33ac4e95113203cfe06a7c466988",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Display Listing Project"
                    }
                }
            }
        }
    }
}